requirements.txt
praw==7.7.1
openai==1.84.0
python-dotenv==1.0.0